Update to newest SoSy-Lab Common and use feature of "annotated" configuration values for replacing the ":: in filenames" hack of RestartAlgorithm and ParallelAlgorithm, which does not work on Windows.
Fixes #273.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@23261 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/value/AbstractExpressionValueVisitor.java
@@ -1429,7 +1429,7 @@ public abstract class AbstractExpressionValueVisitor
       return createSymbolicExpression(valueObject, operandType, unaryOperator, expressionType);
 
     } else {
-      logger.logf(Level.FINE, errorMsg);
+      logger.log(Level.FINE, errorMsg);
       return UnknownValue.getInstance();
     }
   }
